Move the accounting handoff into its own partial
The retry loop pushed OrderManagementSystem.cs 18 lines past its
file-size cap. Raising the baseline is the wrong answer — that guard
exists to force decomposition, and review on the previous PR in this
series was explicit that caps must not be lifted to absorb growth,
least of all growth this branch introduced.

The class is already split across partials, so the handoff replay and
retention methods move to OrderManagementSystem.AccountingHandoff.cs
alongside the fill-identity partial. The order-lifecycle file drops to
2004 lines against a 2098 cap, the delivery-boundary logic gets a
dedicated home, and the ratchet reports 82 reclaimable lines where it
previously failed.
